+++ src/main/kotlin/app/morphe/patcher/PatcherConfig.kt
@@ -8,9 +8,11 @@
 
 package app.morphe.patcher
 
+import app.morphe.patcher.dex.BytecodeMode
+import app.morphe.patcher.dex.DexVerifier
+import app.morphe.patcher.dex.NoOpDexVerifier
 import app.morphe.patcher.resource.CpuArchitecture
 import app.morphe.patcher.resource.ResourceMode
-import brut.androlib.Config
 import java.io.File
 import java.util.logging.Logger
 
@@ -22,16 +24,25 @@ import java.util.logging.Logger
  * @param aaptBinaryPath A path to a custom aapt binary.
  * @param frameworkFileDirectory A path to the directory to cache the framework file in.
  * @param useArsclib Whether to use Arsclib for resource decoding and compiling.
+ * @param verifier The output verifier to use.
  */
 class PatcherConfig(
     internal val apkFile: File,
     private val temporaryFilesPath: File = File("morphe-temporary-files"),
-    private val aaptBinaryPath: String? = null,
-    private val frameworkFileDirectory: String? = null,
-    internal val useArsclib: Boolean = true,
+    @Deprecated("apktool support is deprecated") private val aaptBinaryPath: String? = null,
+    @Deprecated("apktool support is deprecated") private val frameworkFileDirectory: String? = null,
+    @Deprecated("apktool support is deprecated") internal val useArsclib: Boolean = true,
     internal val keepArchitectures: Set<CpuArchitecture> = emptySet(),
+    internal val useBytecodeMode: BytecodeMode = BytecodeMode.STRIP_SAFE,
+    internal val verifier: DexVerifier = NoOpDexVerifier,
 ) {
     private val logger = Logger.getLogger(PatcherConfig::class.java.name)
+
+    init {
+        if (!useArsclib) {
+            logger.warning { "apktool support is deprecated. arsclib will be used." }
+        }
+    }
 
     /**
      * The mode to use for resource decoding and compiling.
@@ -41,18 +52,16 @@ class PatcherConfig(
     internal var resourceMode = ResourceMode.NONE
 
     /**
+     * The mode to use for resource decoding and compiling.
+     *
+     * @see BytecodeMode
+     */
+    internal var bytecodeMode = BytecodeMode.NONE
+
+    /**
      * The path to the temporary apk files directory.
      */
     internal val apkFiles = temporaryFilesPath.resolve("apk")
-
-    /**
-     * The configuration for decoding and compiling resources.
-     */
-    internal val resourceConfig: Config = Config().apply {
-        aaptVersion = 2
-        aaptBinaryPath?.let { setAaptBinaryPath(it) }
-        frameworkFileDirectory?.let { frameworkDirectory = it }
-    }
 
     /**
      * The path to the temporary patched files directory.