+++ src/main/kotlin/app/morphe/patcher/apk/ApkUtils.kt
@@ -48,22 +48,17 @@ object ApkUtils {
      * Applies the [PatcherResult] to the given [apkFile].
      *
      * The order of operation is as follows:
-     * 1. Write patched dex files.
-     * 2. Delete all resources in the target APK
-     * 3. Merge resources.apk compiled by AAPT.
-     * 4. Write raw resources.
-     * 5. Delete resources staged for deletion.
+     * 1. Delete all resources in the target APK.
+     * 2. Merge resources.apk compiled by AAPT.
+     * 3. Write raw resources.
+     * 4. Delete resources staged for deletion.
+     * 5. Write patched dex files.
      * 6. Realign the APK.
      *
      * @param apkFile The file to apply the patched files to.
      */
     fun PatcherResult.applyTo(apkFile: File) {
         ZFile.openReadWrite(apkFile, zFileOptions).use { targetApkZFile ->
-            dexFiles.forEach { dexFile ->
-                targetApkZFile.add(dexFile.name, dexFile.stream)
-                dexFile.stream.close()
-            }
-
             resources?.let { resources ->
                 // Add resources compiled by AAPT.
                 resources.resourcesApk?.let { resourcesApk ->
@@ -76,7 +71,10 @@ object ApkUtils {
                             entry.centralDirectoryHeader.name.startsWith(RES_PREFIX)
                         }.forEach(StoredEntry::delete)
 
-                        targetApkZFile.mergeFrom(resourcesApkZFile) { false }
+                        targetApkZFile.mergeFrom(resourcesApkZFile) { entry ->
+                            // Filter any dex files in case they were packaged inside resources.apk for some reason.
+                            entry.startsWith("classes") && entry.endsWith(".dex")
+                        }
                     }
                 }
 
@@ -93,6 +91,12 @@ object ApkUtils {
                         entry.centralDirectoryHeader.name in resources.deleteResources
                     }.forEach(StoredEntry::delete)
                 }
+            }
+
+            // Run this after resource updates to ensure our dex files don't get overwritten.
+            dexFiles.forEach { dexFile ->
+                targetApkZFile.add(dexFile.name, dexFile.stream)
+                dexFile.stream.close()
             }
 
             logger.info("Aligning APK")
@@ -185,8 +189,8 @@ object ApkUtils {
     class KeyStoreDetails(
         val keyStore: File,
         val keyStorePassword: String? = null,
-        val alias: String = "Morphe Key",
-        val password: String = "",
+        val alias: String,
+        val password: String,
     )
 
     /**